* master:
  Best models table (#253)
  Bugfix for reattempt_failures=False (#262)
  issue fix (#265)
  Staging recovery bugfix (#259)

# Conflicts:
#	docs/more_info/changelog.rst
#	dynamite/model.py

+++ dynamite/model.py
@@ -172,12 +172,21 @@ class AllModels(object):
                                  [d[:d[:-1].rindex('/')+1]
                                  for d in self.table[to_delete]['directory']]
                                 )
-            self.logger.info(f'Will remove {len(dirs_to_delete)} '
+            self.logger.info(f'Will try to remove {len(dirs_to_delete)} '
                              'unique orblibs.')
             for directory in dirs_to_delete:
                 try:
-                    shutil.rmtree(directory)
-                    self.logger.info(f'Model directory {directory} removed.')
+                    # Only remove orblib directories that are not used by
+                    # already completed models
+                    orblibs_keep = [d[:d[:-1].rindex('/')+1] for d in
+                     self.table[np.where(self.table['all_done'])]['directory']]
+                    if directory in set(orblibs_keep):
+                        self.logger.info(f'Orblib directory {directory} in '
+                                         'use by existing model - untouched.')
+                    else:
+                        shutil.rmtree(directory)
+                        self.logger.info(f'Orblib directory {directory} '
+                                         'removed.')
                 except:
                     self.logger.warning(f'Cannot remove orblib in {directory},'
                         ' perhaps it has already been removed before.')
@@ -525,7 +534,7 @@ class AllModels(object):
             The default is None.
         delta : float, optional
             The threshold value. Models with chi2 values differing
-            from the opimum by at most delta will be returned. If none,
+            from the opimum by at most delta will be returned. If None,
             models within 10% of the optimal value will be returned. If
             delta is negative, models that are NOT within a delta
             threshold of the best are returned.
@@ -552,6 +561,75 @@ class AllModels(object):
         else:
             models = self.table[self.table[which_chi2] > chi2_min-delta]
         return models
+
+    def make_best_models_table(self,
+                               which_chi2=None,
+                               n=None,
+                               delta=None,
+                               filename=None):
+        """Make a table of the best models and save it to disk
+
+        Parameters
+        ----------
+        which_chi2 : str, optional
+            Which chi2 is used for determining the best models. If None, the
+            setting from the configuration file will be used.
+            The default is None.
+        n : int, optional
+            How many models to get. If None, n will be ignored.
+            Default: if delta is specified, the default is none; if delta
+            is None, the default is 10.
+        delta : float, optional
+            The threshold value. Models with chi2 values differing
+            from the opimum by at most delta will be returned. If None,
+            delta will be ignored. The default is None.
+        filename : str, optional
+            File name of the best models table. The file is written into the
+            output directory specified in the config file. If None, the name
+            is the same as for the all models table but with '_best' added
+            to the base file name. If the file already exists, a warning
+            is logged, the existing file is backed up ('_backup' added),
+            and then overwritten. The default is None.
+
+        Raises
+        ------
+        ValueError
+            If both n and delta are specified (i.e., both are not None).
+
+        Returns
+        -------
+        int
+            The number of models in the best models table.
+
+        """
+        if n is not None and delta is not None:
+            text = 'Cannot specify both n and delta - choose one...'
+            self.logger.error(text)
+            raise ValueError(text)
+        elif n is None and delta is None:
+            n = 10
+            self.logger.info('No parameters specified - making table with '
+                             '10 best models')
+        if filename is None:
+            path_noext, ext = os.path.splitext(self.filename)
+            filename = path_noext + '_best' + ext
+        else:
+            filename = self.config.settings.io_settings['output_directory'] + \
+                       filename
+        if os.path.exists(filename):
+            path_noext, ext = os.path.splitext(filename)
+            backup_filename = path_noext + '_backup' + ext
+            shutil.copy2(filename, backup_filename)
+            self.logger.warning(f'File {filename} will be overwritten, '
+                                f'backup {backup_filename} created.')
+        if n is not None:
+            table_best = self.get_best_n_models(which_chi2=which_chi2, n=n)
+        else:
+            table_best=self.get_mods_within_chi2_thresh(which_chi2=which_chi2,
+                                                        delta=delta)
+        table_best.write(filename, format='ascii.ecsv', overwrite=True)
+        self.logger.info(f'Table of best models written to file {filename}.')
+        return len(table_best)
 
     def remove_unused_orblibs(self):
         """
